+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All Notable changes to `flipboxdigital/relay` will be documented in this file
 
+## Unreleased
+### Changed
+- RelayHelper::createResolver accepts a callable as valid middleware
+
 ## 2.2.0 - 2017-01-05
 ### Removed
 - Middleware configurable response success codes as they are overreaching